// autogpt_platform/frontend/src/app/api/__generated__/models/libraryFolder.ts
/**
 * Generated by orval v7.13.0 🍺
 * Do not edit manually.
 * AutoGPT Agent Server
 * This server is used to execute agents that are created by the AutoGPT system.
 * OpenAPI spec version: 0.1
 */
import type { LibraryFolderIcon } from "./libraryFolderIcon";
import type { LibraryFolderColor } from "./libraryFolderColor";
import type { LibraryFolderParentId } from "./libraryFolderParentId";

/**
 * Represents a folder for organizing library agents.
 */
export interface LibraryFolder {
  id: string;
  user_id: string;
  name: string;
  icon?: LibraryFolderIcon;
  color?: LibraryFolderColor;
  parent_id?: LibraryFolderParentId;
  created_at: Date;
  updated_at: Date;
  agent_count?: number;
  subfolder_count?: number;
}